[#12297] refactor(secret): Keep secret params out of catalog plugins
Secrets are resolved in dispatchers into property URNs; connector
SupportsSchemas and catalog ops only take the final properties map.
