+++ catalogs/catalog-jdbc-common/src/main/java/org/apache/gravitino/catalog/jdbc/JdbcCatalogOperations.java
@@ -75,8 +75,6 @@ import org.apache.gravitino.rel.expressions.distributions.Distribution;
 import org.apache.gravitino.rel.expressions.sorts.SortOrder;
 import org.apache.gravitino.rel.expressions.transforms.Transform;
 import org.apache.gravitino.rel.indexes.Index;
-import org.apache.gravitino.secret.SecretBinding;
-import org.apache.gravitino.secret.SecretReference;
 import org.apache.gravitino.utils.MapUtils;
 import org.apache.gravitino.utils.PrincipalUtils;
 import org.slf4j.Logger;
@@ -265,29 +263,6 @@ public class JdbcCatalogOperations implements CatalogOperations, SupportsSchemas
   @Override
   public JdbcSchema createSchema(
       NameIdentifier ident, String comment, Map<String, String> properties)
-      throws NoSuchCatalogException, SchemaAlreadyExistsException {
-    return createSchema(ident, comment, properties, Collections.emptyMap(), Collections.emptyMap());
-  }
-
-  /**
-   * Creates a new schema with the provided identifier, comment, metadata, and secret options.
-   *
-   * @param ident The identifier of the schema to create.
-   * @param comment The comment for the schema.
-   * @param properties The properties for the schema.
-   * @param secretBindings The secret bindings for the schema.
-   * @param secretReferences The secret references for the schema.
-   * @return The created {@link JdbcSchema}.
-   * @throws NoSuchCatalogException If the provided namespace is invalid or does not exist.
-   * @throws SchemaAlreadyExistsException If a schema with the same name already exists.
-   */
-  @Override
-  public JdbcSchema createSchema(
-      NameIdentifier ident,
-      String comment,
-      Map<String, String> properties,
-      Map<String, SecretBinding> secretBindings,
-      Map<String, SecretReference> secretReferences)
       throws NoSuchCatalogException, SchemaAlreadyExistsException {
 
     StringIdentifier identifier = StringIdentifier.fromProperties(properties);
