+++ catalogs/catalog-lakehouse-paimon/src/main/java/org/apache/gravitino/catalog/lakehouse/paimon/PaimonCatalogOperations.java
@@ -29,7 +29,6 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.Maps;
 import java.time.Instant;
 import java.util.Arrays;
-import java.util.Collections;
 import java.util.List;
 import java.util.Map;
 import java.util.Optional;
@@ -71,8 +70,6 @@ import org.apache.gravitino.rel.expressions.distributions.Strategy;
 import org.apache.gravitino.rel.expressions.sorts.SortOrder;
 import org.apache.gravitino.rel.expressions.transforms.Transform;
 import org.apache.gravitino.rel.indexes.Index;
-import org.apache.gravitino.secret.SecretBinding;
-import org.apache.gravitino.secret.SecretReference;
 import org.apache.gravitino.utils.MapUtils;
 import org.apache.gravitino.utils.PrincipalUtils;
 import org.apache.paimon.catalog.Catalog;
@@ -186,30 +183,6 @@ public class PaimonCatalogOperations
   @Override
   public PaimonSchema createSchema(
       NameIdentifier identifier, String comment, Map<String, String> properties)
-      throws NoSuchCatalogException, SchemaAlreadyExistsException {
-    return createSchema(
-        identifier, comment, properties, Collections.emptyMap(), Collections.emptyMap());
-  }
-
-  /**
-   * Creates a new schema with the provided identifier, comment, metadata, and secret options.
-   *
-   * @param identifier The identifier of the schema to create.
-   * @param comment The comment for the new schema.
-   * @param properties The properties for the new schema.
-   * @param secretBindings The secret bindings for the schema.
-   * @param secretReferences The secret references for the schema.
-   * @return The newly created {@link PaimonSchema} instance.
-   * @throws NoSuchCatalogException If the provided namespace is invalid or does not exist.
-   * @throws SchemaAlreadyExistsException If a schema with the same name already exists.
-   */
-  @Override
-  public PaimonSchema createSchema(
-      NameIdentifier identifier,
-      String comment,
-      Map<String, String> properties,
-      Map<String, SecretBinding> secretBindings,
-      Map<String, SecretReference> secretReferences)
       throws NoSuchCatalogException, SchemaAlreadyExistsException {
     String currentUser = currentUser();
     PaimonSchema createdSchema =
