+++ core/src/main/java/org/apache/gravitino/catalog/ManagedSchemaOperations.java
@@ -42,8 +42,6 @@ import org.apache.gravitino.exceptions.NonEmptySchemaException;
 import org.apache.gravitino.exceptions.SchemaAlreadyExistsException;
 import org.apache.gravitino.meta.AuditInfo;
 import org.apache.gravitino.meta.SchemaEntity;
-import org.apache.gravitino.secret.SecretBinding;
-import org.apache.gravitino.secret.SecretReference;
 import org.apache.gravitino.utils.PrincipalUtils;
 
 public abstract class ManagedSchemaOperations implements SupportsSchemas {
@@ -91,12 +89,7 @@ public abstract class ManagedSchemaOperations implements SupportsSchemas {
   }
 
   @Override
-  public Schema createSchema(
-      NameIdentifier ident,
-      String comment,
-      Map<String, String> properties,
-      Map<String, SecretBinding> secretBindings,
-      Map<String, SecretReference> secretReferences)
+  public Schema createSchema(NameIdentifier ident, String comment, Map<String, String> properties)
       throws NoSuchCatalogException, SchemaAlreadyExistsException {
     try {
       if (store().exists(ident, Entity.EntityType.SCHEMA)) {
